examples/python/clients/httpx/pyproject.toml
[project]
name = "x402-httpx-example"
version = "0.1.0"
description = "Example of using x402 v2 SDK with httpx"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "x402[httpx,evm,svm,avm]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
package = false

[tool.uv.sources]
x402 = { path = "../../../../python/x402", editable = true }
